Override the default Flask and JWT error handlers
Let any raised exceptions be caught and handled by the exception
handlers defined in exceptions.py. Also subclass the Flask and Request
classes so get_json can be overridden and changed to raise an exception
if the request MIME type is not set to JSON. This removes the
requirement to check for MIME type before parsing the request as JSON.
If the request is not JSON the raised exception will be handled by the
new exception handlers.